Speeded up the calculation of z.
・ユニットの入力の総和の計算を高速化した。
