+++ doc/MNISTLoader.html
@@ -94,7 +94,7 @@
   
   <dl>
     <dt>Defined in:</dt>
-    <dd>mnist_loader.rb</dd>
+    <dd>.\mnist_loader.rb</dd>
   </dl>
   
 </div>
@@ -417,7 +417,7 @@
 15</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'mnist_loader.rb', line 12</span>
+      <pre class="code"><span class="info file"># File '.\mnist_loader.rb', line 12</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_initialize'>initialize</span><span class='lparen'>(</span><span class='id identifier rubyid_image_path'>image_path</span><span class='comma'>,</span> <span class='id identifier rubyid_label_path'>label_path</span><span class='rparen'>)</span>
   <span class='ivar'>@image_path</span> <span class='op'>=</span> <span class='id identifier rubyid_image_path'>image_path</span>
@@ -522,7 +522,7 @@
 56</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'mnist_loader.rb', line 52</span>
+      <pre class="code"><span class="info file"># File '.\mnist_loader.rb', line 52</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_binarize'>binarize</span><span class='lparen'>(</span><span class='id identifier rubyid_inputs'>inputs</span><span class='rparen'>)</span>
   <span class='kw'>return</span> <span class='id identifier rubyid_inputs'>inputs</span><span class='period'>.</span><span class='id identifier rubyid_map'>map</span><span class='lbrace'>{</span><span class='op'>|</span><span class='id identifier rubyid_pixel'>pixel</span><span class='op'>|</span>
@@ -603,7 +603,7 @@
 36</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'mnist_loader.rb', line 32</span>
+      <pre class="code"><span class="info file"># File '.\mnist_loader.rb', line 32</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_load_gzipped_idx_file'>load_gzipped_idx_file</span><span class='lparen'>(</span><span class='id identifier rubyid_path'>path</span><span class='rparen'>)</span>
   <span class='id identifier rubyid_file'>file</span> <span class='op'>=</span> <span class='const'>File</span><span class='period'>.</span><span class='id identifier rubyid_open'>open</span><span class='lparen'>(</span><span class='id identifier rubyid_path'>path</span><span class='comma'>,</span><span class='tstring'><span class='tstring_beg'>&quot;</span><span class='tstring_content'>rb</span><span class='tstring_end'>&quot;</span></span><span class='rparen'>)</span>
@@ -662,7 +662,7 @@
 21</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'mnist_loader.rb', line 19</span>
+      <pre class="code"><span class="info file"># File '.\mnist_loader.rb', line 19</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_load_images'>load_images</span>
   <span class='ivar'>@images</span> <span class='op'>=</span> <span class='id identifier rubyid_load_gzipped_idx_file'>load_gzipped_idx_file</span><span class='lparen'>(</span><span class='ivar'>@image_path</span><span class='rparen'>)</span>
@@ -719,7 +719,7 @@
 27</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'mnist_loader.rb', line 25</span>
+      <pre class="code"><span class="info file"># File '.\mnist_loader.rb', line 25</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_load_labels'>load_labels</span>
   <span class='ivar'>@labels</span> <span class='op'>=</span> <span class='id identifier rubyid_load_gzipped_idx_file'>load_gzipped_idx_file</span><span class='lparen'>(</span><span class='ivar'>@label_path</span><span class='rparen'>)</span>
@@ -816,7 +816,7 @@
 46</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'mnist_loader.rb', line 42</span>
+      <pre class="code"><span class="info file"># File '.\mnist_loader.rb', line 42</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_normalize'>normalize</span><span class='lparen'>(</span><span class='id identifier rubyid_inputs'>inputs</span><span class='rparen'>)</span>
   <span class='kw'>return</span> <span class='id identifier rubyid_inputs'>inputs</span><span class='period'>.</span><span class='id identifier rubyid_map'>map</span><span class='lbrace'>{</span><span class='op'>|</span><span class='id identifier rubyid_pixel'>pixel</span><span class='op'>|</span>
@@ -901,7 +901,7 @@
 69</pre>
     </td>
     <td>
-      <pre class="code"><span class="info file"># File 'mnist_loader.rb', line 61</span>
+      <pre class="code"><span class="info file"># File '.\mnist_loader.rb', line 61</span>
 
 <span class='kw'>def</span> <span class='id identifier rubyid_print_ascii'>print_ascii</span><span class='lparen'>(</span><span class='id identifier rubyid_inputs'>inputs</span><span class='rparen'>)</span>
   <span class='id identifier rubyid_inputs'>inputs</span> <span class='op'>=</span> <span class='id identifier rubyid_inputs'>inputs</span><span class='period'>.</span><span class='id identifier rubyid_map'>map</span> <span class='lbrace'>{</span><span class='op'>|</span><span class='id identifier rubyid_pixel'>pixel</span><span class='op'>|</span> <span class='id identifier rubyid_pixel'>pixel</span><span class='op'>*</span><span class='int'>255</span><span class='rbrace'>}</span>
@@ -922,7 +922,7 @@
 </div>
 
       <div id="footer">
-  Generated on Sun Oct 25 04:56:22 2020 by
+  Generated on Wed Oct 28 17:54:35 2020 by
   <a href="http://yardoc.org" title="Yay! A Ruby Documentation Tool" target="_parent">yard</a>
   0.9.25 (ruby-2.6.3).
 </div>
