+++ doc/top-level-namespace.html
@@ -123,7 +123,7 @@
 </div>
 
       <div id="footer">
-  Generated on Sun Oct 25 04:56:21 2020 by
+  Generated on Wed Oct 28 17:54:34 2020 by
   <a href="http://yardoc.org" title="Yay! A Ruby Documentation Tool" target="_parent">yard</a>
   0.9.25 (ruby-2.6.3).
 </div>
