+++ examples/NN/xor/neural__network_00003aT0_00003a_00003alayer0_00003aT0_00003a_00003afunc0_00003aT1_00003a_00003amy__lut_00003aT2.v
@@ -28,35 +28,35 @@ module neural__network_00003aT0_00003a_00003alayer0_00003aT0_00003a_00003afunc0_
 
       lut[32'd0] = $signed(32'd0);
 
-      lut[32'd1] = $signed(32'd12);
+      lut[32'd1] = $signed(32'd16);
 
-      lut[32'd2] = $signed(32'd15);
+      lut[32'd2] = $signed(32'd32);
 
-      lut[32'd3] = $signed(32'd15);
+      lut[32'd3] = $signed(32'd48);
 
-      lut[32'd4] = $signed(32'd15);
+      lut[32'd4] = $signed(32'd64);
 
-      lut[32'd5] = $signed(32'd15);
+      lut[32'd5] = $signed(32'd80);
 
-      lut[32'd6] = $signed(32'd15);
+      lut[32'd6] = $signed(32'd96);
 
-      lut[32'd7] = $signed(32'd15);
+      lut[32'd7] = $signed(32'd112);
 
-      lut[32'd8] = $signed(-32'd15);
+      lut[32'd8] = $signed(32'd0);
 
-      lut[32'd9] = $signed(-32'd15);
+      lut[32'd9] = $signed(32'd0);
 
-      lut[32'd10] = $signed(-32'd15);
+      lut[32'd10] = $signed(32'd0);
 
-      lut[32'd11] = $signed(-32'd15);
+      lut[32'd11] = $signed(32'd0);
 
-      lut[32'd12] = $signed(-32'd15);
+      lut[32'd12] = $signed(32'd0);
 
-      lut[32'd13] = $signed(-32'd15);
+      lut[32'd13] = $signed(32'd0);
 
-      lut[32'd14] = $signed(-32'd15);
+      lut[32'd14] = $signed(32'd0);
 
-      lut[32'd15] = $signed(-32'd12);
+      lut[32'd15] = $signed(32'd0);
 
    end
