+++ lib/fast_neurons.rb
@@ -10,7 +10,7 @@ ALPHA = 1.6732632423543772848170429916717
 # FastNeurons is a simple and fast library using NMatrix for building neural networks.<br>
 # Currently, it supports fully connected neural network and restricted boltzmann machine.<br>
 # More models will be added gradually.<br>
-# @version 1.7.0
+# @version 1.8.0
 # @since 1.0.0
 # @author Ryota Sakai, Yusuke Tomimoto
 module FastNeurons
