+++ lib/models/NN.rb
@@ -1,32 +1,3 @@
-require 'HDLRuby'
-require 'fileutils'
-require 'HDLRuby'
-require 'HDLRuby/hruby_check.rb'
-require 'HDLRuby/hruby_low2high'
-require 'HDLRuby/hruby_low2c'
-require 'HDLRuby/hruby_low2vhd'
-require 'HDLRuby/hruby_low_fix_types'
-require 'HDLRuby/hruby_low_without_outread'
-require 'HDLRuby/hruby_low_with_bool'
-require 'HDLRuby/hruby_low_bool2select'
-require 'HDLRuby/hruby_low_without_select'
-require 'HDLRuby/hruby_low_without_namespace'
-require 'HDLRuby/hruby_low_without_bit2vector'
-require 'HDLRuby/hruby_low_with_port'
-require 'HDLRuby/hruby_low_with_var'
-require 'HDLRuby/hruby_low_without_concat'
-require 'HDLRuby/hruby_low_without_connection'
-require 'HDLRuby/hruby_low_cleanup'
-require 'HDLRuby/hruby_verilog.rb'
-require 'HDLRuby/backend/hruby_allocator'
-require 'HDLRuby/backend/hruby_c_allocator'
-
-configure_high
-require_relative 'modules/network_constructor.rb'
-
-# A proc object for instantiating module written by HDLRuby.
-$network_constructor_caller = proc{ |*args, &blk| network_constructor(*args, &blk) }
-
 module FastNeurons
   # Describes a standard fully connected NN based on backpropagation.
   # @example learning of xor
@@ -139,7 +110,7 @@ module FastNeurons
       @idm_geometry = @weights_geometry.clone
 
       # Create identity matrix.
-      @idm = @idm_geometry.map{ |g| NMatrix.eye([g[0],g[0]]) }
+      @idm = @idm_geometry.map{ |g| NMatrix.eye([g[0],g[0]]) }            
 
       # Set the coefficients of derivatives.
       @coefficients = NMatrix.ones_like(@a[@neuron_columns.size])
@@ -381,9 +352,17 @@ module FastNeurons
     # z = inputs * weights + biases
     # @param [Integer] row the number of layer currently computing
     # @since 1.0.0
-    def compute_z(row)
-      # Compute the values before the activation function is applied.      
-      @z[row] = NMatrix::BLAS.gemm(@weights[row], @a[row], @biases[row].clone, 1.0, 1.0)      
+    def compute_z(row)      
+      # BLAS.gemm performs the following calculations.
+      #   C = (alpha * A * B) + (beta * C)
+      # In this case, the calculation results are stored in Matrix C.
+      # For this reason, need to duplicate the @biases[row] value in @z[row] in advance.
+
+      # Duplicate @biases[row] value to @z[row].
+      @z[row] = NMatrix::BLAS.gemm(@idm[row], @biases[row])
+      
+      # Compute the values before the activation function is applied.  
+      @z[row] = NMatrix::BLAS.gemm(@weights[row], @a[row], @z[row], 1.0, 1.0)                        
     end
 
     # Compute neurons statuses.
@@ -430,8 +409,8 @@ module FastNeurons
         @delta[@neuron_columns.size-1] = @loss_derivative.call(@T, @a[-1]) * @derivative_activation_function[@neuron_columns.size-1] * @coefficients
       end
 
-      # Compute loss derivatives.
-      @loss_derivative_weights[@neuron_columns.size-1] += NMatrix::BLAS.gemm(@delta[@neuron_columns.size-1], @a[@neuron_columns.size-1].transpose)
+      # Compute loss derivatives.            
+      @loss_derivative_weights[@neuron_columns.size-1] += NMatrix::BLAS.gemm(@delta[@neuron_columns.size-1], @a[@neuron_columns.size-1].transpose)      
       @loss_derivative_biases[@neuron_columns.size-1] += @delta[@neuron_columns.size-1]
 
       (@neuron_columns.size-2).downto(0) do |i|
@@ -472,8 +451,8 @@ module FastNeurons
     # Compute derivatives of weights.
     # @param [Integer] row the number of layer currently computing
     # @since 1.2.0
-    def compute_weights_derivatives(row)
-      @loss_derivative_weights[row] += NMatrix::BLAS.gemm(@delta[row], @a[row].transpose)
+    def compute_weights_derivatives(row)                  
+      @loss_derivative_weights[row] += NMatrix::BLAS.gemm(@delta[row], @a[row].transpose)      
     end
 
     # Compute derivatives of biases.
@@ -500,15 +479,15 @@ module FastNeurons
     # @since 1.0.0
     def update_weights(row)
       @loss_derivative_weights[row] = @loss_derivative_weights[row] / @batch_size.to_f
-      @weights[row] = NMatrix::BLAS.gemm(@idm[row], @loss_derivative_weights[row], @weights[row].clone, @learning_rate, 1.0)
+      @weights[row] = NMatrix::BLAS.gemm(@idm[row], @loss_derivative_weights[row], @weights[row], @learning_rate, 1.0)
     end
 
     # Update biases.
     # @param [Integer] row the number of layer currently computing
     # @since 1.0.0
     def update_biases(row)
-      @loss_derivative_biases[row] = @loss_derivative_biases[row] / @batch_size.to_f
-      @biases[row] = NMatrix::BLAS.gemm(@idm[row], @loss_derivative_biases[row], @biases[row].clone, @learning_rate, 1.0)
+      @loss_derivative_biases[row] = @loss_derivative_biases[row] / @batch_size.to_f      
+      @biases[row] = NMatrix::BLAS.gemm(@idm[row], @loss_derivative_biases[row], @biases[row], @learning_rate, 1.0)            
     end
 
     # Update biases and weights.
@@ -523,9 +502,15 @@ module FastNeurons
     # Compute loss by loss function.
     # @since 1.5.0
     def compute_loss
-      @loss = @loss_antiderivative.call(@T, @a[-1])
+      @loss += @loss_antiderivative.call(@T, @a[-1])      
+    end
+    
+    # Initialize and output loss by loss function.
+    # @since 1.8.0
+    def initialize_loss
       puts "loss : #{@loss}"
-    end    
+      @loss = 0.0
+    end
 
     # Get outputs of the layer of neural network.
     # @param [Integer] row the row number you want to get outputs
@@ -643,76 +628,6 @@ module FastNeurons
           @weights.push(NMatrix.new(@weights_geometry[i], weights_array))
         end
         puts "#{@weights}"
-      end
-    end
-
-    # Instantiate neural network module written by HDLRuby.
-    # @param [Array or proc] functions an activation function of each layer
-    # @param [Integer] integer_width the width of integer part of fixed point
-    # @param [Integer] decimal_width the width of decimal part of fixed point
-    # @param [Integer] address_width the width of address for Look up table of activation functions
-    # @param [Array] inputs inputs of neural network(for simulation)
-    # @param [Array] weights an array of all weights of neural network
-    # @param [Array] biases an array of all biases of neural network
-    # @since 1.4.0
-    def instantiate(functions, integer_width, decimal_width, address_width, inputs, weights = nil, biases = nil)
-      @functions = functions
-      @integer_width = integer_width
-      @decimal_width = decimal_width
-      @address_width = address_width
-
-      @types = signed[@integer_width, @decimal_width]
-
-      @inputs = inputs
-      @module_weights = weights.nil? ? @weights.map.with_index{ |w, i| w.to_a.flatten.each_slice(@weights_geometry[i][1]).to_a } : weights
-      @module_biases = biases.nil? ? @biases.map{ |b| b.to_a.flatten } : biases
-
-
-      @instance =  $network_constructor_caller.(@columns, @functions, @types, @integer_width, @decimal_width, @address_width, @inputs, @module_weights, @module_biases).(:neural_network)
-    end
-
-    # Generate the Verilog files of neural network module.
-    # @param [String] folder_name output folder's name
-    # @since 1.4.0
-    def to_verilog(folder_name)
-      # Generate the low level representation.
-      top_system = @instance.to_low.systemT
-
-      top_system.each_systemT_deep do |systemT|
-        systemT.to_upper_space!
-        systemT.to_global_systemTs!
-        systemT.initial_concat_to_timed!
-        systemT.with_port!
-      end
-
-      output = folder_name
-      basename = output + "/neural_network"
-
-      # Create a directory if necessary.
-      unless File.directory?(output)
-        FileUtils.mkdir_p(output)
-      end
-
-      # Prepare the initial name for the main file.
-      name = basename + ".v"
-      # Multiple files generation mode.
-      top_system.each_systemT_deep do |systemT|
-        # Generate the name if necessary.
-        unless name
-          name = output + "/" + HDLRuby::Verilog.name_to_verilog(systemT.name) + ".v"
-        end
-        
-        # Open the file for current systemT
-        outfile = File.open(name,"w")
-    
-        # Generate the Verilog code in to.    
-        outfile << systemT.to_verilog
-
-        # Close the file.
-        outfile.close
-
-        # Clears the name.
-        name = nil
       end
     end
   end
